数值验证

+++ adwpdownload.php
@@ -27,8 +27,8 @@ try{
 	require_once 'tasks/statis/ReqStatis.class.php';
 	$reqStatis = new ReqStatis();
 	$cpid = isset($_GET['cpid'])?$_GET['cpid']:'';
-	$type = isset($_GET['type'])?$_GET['type']:0;
-	$channel = isset($_GET['channel'])?$_GET['channel']:0;
+	$type = (int)(isset($_GET['type'])?$_GET['type']:0);
+	$channel = (int)(isset($_GET['channel'])?$_GET['channel']:0);
 	$height = 0;
 	$width  = 0;
  	$reqStatis->recordDownloadRequest($id, COOLXIU_TYPE_ANDROIDESK_WALLPAPER, $height, $width, $cpid, $url, $type, $channel);
