+++ hotword.php
@@ -5,7 +5,7 @@
  *兼容了安卓壁纸的资源
  */
 
-$nCoolType = isset($_GET['type'])?$_GET['type']:0;  //cooltype:主题、壁纸、铃声、专题等分类
+$nCoolType = (int)(isset($_GET['type'])?$_GET['type']:0);  //cooltype:主题、壁纸、铃声、专题等分类
 $bColor    = isset($_GET['iscolor'])?$_GET['iscolor']:false;  //iscolor:获取延时关键词
 require_once 'tasks/CoolShow/CoolShowSearch.class.php';
