+++ test.php
@@ -1,6 +1,6 @@
 <?php
-$req_page = isset($_GET['page'])?$_GET['page']:0;
-$req_num  = isset($_GET['reqNum'])?$_GET['reqNum']:10;
+$req_page = (int)(isset($_GET['page'])?$_GET['page']:0);
+$req_num  = (int)(isset($_GET['reqNum'])?$_GET['reqNum']:10);
 $start 	  = $req_num * $req_page;
 
 $nCoolType   = (int)(isset($_GET['type']))?$_GET['type']:0;
