+++ webinterface.php
@@ -1,8 +1,8 @@
 <?php
 
 
-$req_page = isset($_GET['page'])?$_GET['page']:0;
-$req_num  = isset($_GET['reqNum'])?$_GET['reqNum']:10;
+$req_page = (int)(isset($_GET['page'])?$_GET['page']:0);
+$req_num  = (int)(isset($_GET['reqNum'])?$_GET['reqNum']:10);
 $start 	  = $req_num * $req_page;
 
 require_once 'lib/WriteLog.lib.php';
