+++ weblucene.php
@@ -4,8 +4,8 @@ require_once 'tasks/lucene/luceneTask.php';
 
 $nCoolType   = isset($_GET['type'])?$_GET['type']:-1;
 $keyWord  	 = isset($_GET['keyword'])?$_GET['keyword']:'';
-$nPage     	 = isset($_GET['page'])?$_GET['page']:0;
-$nLimit      = isset($_GET['reqNum'])?$_GET['reqNum']:100;
+$nPage     	 = (int)(isset($_GET['page'])?$_GET['page']:0);
+$nLimit      = (int)(isset($_GET['reqNum'])?$_GET['reqNum']:100);
 
 if (empty($keyWord)){
 	echo get_rsp_result(false, 'keyword is null');
